pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
balanced_wrapping = true
default_section = "THIRDPARTY"
force_grid_wrap = 0
include_trailing_comma = true
known_first_party = "dash_charts"
length_sort = false
line_length = 120
multi_line_output = 5

[tool.commitizen]
name = "cz_legacy"
change_type_order = [
    "BREAKING CHANGE",
    "Feat",
    "Fix",
    "Refactor",
    "Perf",
    "New (Old)",
    "Change (Old)",
    "Fix (Old)",
]
version = "0.1.0a1"
version_files = ["pyproject.toml", "dash_charts/__init__.py", "appveyor.yml"]

[tool.commitizen.cz_legacy_map]
Chg = "Change (Old)"
Fix = "Fix (Old)"
New = "New (Old)"

[tool.poetry]
name = "dash_charts"
version = "0.1.0a1"
description = "Python package for Plotly/Dash apps with support for multi-page, modules, and new charts such as Pareto with an Object Orient Approach"
license = "MIT"
authors = ["Kyle King <dev.act.kyle@gmail.com>"]
maintainers = []
repository = "https://github.com/kyleking/dash_charts"
documentation = "https://github.com/kyleking/dash_charts/docs"
readme = "docs/README.md"
include = ["LICENSE.md"]
keywords = ["plotly-dash", "plotly-python"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Web Environment",
    "Framework :: Dash",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Topic :: Database :: Front-Ends",
    "Topic :: Scientific/Engineering :: Visualization",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
] # https://pypi.org/classifiers/
# And based on: https://github.com/plotly/dash/blob/6cfb7874800152794d8d603e8d9c4334bf61e3fd/setup.py#L47-L70

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/kyleking/dash_charts/issues"

[tool.poetry.dependencies]
python = "^3.8"
attrs-strict = ">=0.2.2"
calcipy = "^0.10.0"
cerberus = "*"
# dash = ">=2.0.0"
dash-bootstrap-components = ">=1.0.0"
dataset = "*"
dominate = "*"
flask-caching = ">=1.10.1"
implements = "*"
lxml = ">=4.7.1"
markdown = "*"
numpy = ">=1.22.2"
Palettable = "*"
pandas = "*"
psutil = "*"
python-box = "*"
scipy = "*"
tqdm = ">=4.62.3"

# sqlite-utils = "*"
# datasette-vega = "*"
# great-expectations = "*"

# FIXME: Required only for nox_coverage and optional for subset of use cases
astor = ">=0.8.1"
jsonpickle = ">=2.1.0"
beautifulsoup4 = ">=4.10.0"

# FIXME: In-progress testing of a better dash table
dash-tabulator = ">=0.4.2"
dash-extensions = ">=0.0.55"

[tool.poetry.dev-dependencies]
calcipy = { version = "*", extras = [
    "dev",
    "lint",
    "test",
    "commitizen_legacy",
] }

# Experimental dependencies
# archan = { git = "https://github.com/pawamoy/archan.git" }
# dependenpy = "^3.3.0"
# archan-pylint = { git = "https://github.com/pawamoy/archan-pylint" }

# FIXME: testing extras are required for nox-coverage
[tool.poetry.dependencies.dash]
# [tool.poetry.dev-dependencies.dash]
extras = ["testing"]
version = ">=2.0.0"

[tool.poetry.extras]
matplotlib = ["matplotlib"]
